added functions new functions fix betty style
